Made post statuses explicit with a database table and associated module constants. Updated shell scripts to delete images on database drop and reset. Added user confirmation to shell scripts that destroy data.

+++ scripts/dropdb.sh
@@ -1,3 +1,19 @@
 #!/bin/zsh
 
-dropdb skyfilter -i -U admin
+echo "This will delete all existing posts and images in the database. Are you sure you want to proceed? (yes/no) " 
+read answer
+
+case $answer in
+    [Yy]es )
+        echo "Dropping dataset..."
+        dropdb skyfilter -i -U admin
+        find database/images -mindepth 1 -not -name '.gitignore' -exec rm -rf {} +
+        ;;
+    [Nn]o )
+        echo "Operation aborted."
+        exit
+        ;;
+    * )
+        echo "Operation aborted. Please answer 'yes' or 'no'."
+        ;;
+esac
